requirements.txt
numpy
opencv-python
paramiko
python-dotenv
scp
